+++ client/src/config/setsSearchTags.ts
@@ -6,6 +6,12 @@ export const archive: IHalLink = {
   searchTag: 'lux:setItemsWithImages',
 }
 
+// Used to request and display unit data
+export const setWithMemberOf: IHalLink = {
+  title: 'Explore',
+  searchTag: 'lux:objectOrSetMemberOfSet',
+}
+
 // Used for retrieving and rendering the set's unit
 export const setUnit: IHalLink = {
   title: 'Collection',