+++ client/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Changed HAL link for display archive hierarchies ([#458](https://github.com/project-lux/lux-frontend/issues/458)).
+
 ### Removed
 
 ### Fixed